side panel with page names correct
