+++ .idea/workspace.xml
@@ -5,10 +5,14 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="1b691dee-2172-4fb7-95fd-956c3782cad2" name="Changes" comment="Initial commit of streamlit dashboard project">
-      <change afterPath="$PROJECT_DIR$/src/pages/page1.py" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/pages/page2.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/Home.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/job-profile-self-assessment.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/job-profile-self-assessment.iml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/dashboard.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/dashboard.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Dockerfile" beforeDir="false" afterPath="$PROJECT_DIR$/Dockerfile" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/README.md" beforeDir="false" afterPath="$PROJECT_DIR$/README.md" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/dashboard.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/pages/page1.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/1_page_1.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/pages/page2.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/2_page_2.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
