+++ Dockerfile
@@ -14,4 +14,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 EXPOSE 8501
 
 # Run streamlit when the container launches
-CMD ["streamlit", "run", "src/dashboard.py"]
+CMD ["streamlit", "run", "src/Home.py"]
