+++ src/pages/1_page_1.py
@@ -1,0 +1,6 @@
+import streamlit as st
+
+st.set_page_config(page_title="page 1")
+
+st.title("Page 1")
+st.write("This is Page 1 of the dashboard.")
