+++ src/pages/2_page_2.py
@@ -1,0 +1,7 @@
+import streamlit as st
+
+st.set_page_config(page_title="page 2")
+
+
+st.title("Page 2")
+st.write("This is Page 2 of the dashboard.")
