--- src/pages/page1.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-def app():
-    st.title("Page 1")
-    st.write("This is Page 1 of the dashboard.")
