--- src/pages/page2.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-def app():
-    st.title("Page 2")
-    st.write("This is Page 2 of the dashboard.")
